requirements.txt
nvidia-cudnn-cu11==8.6.0.163
tensorflow==2.12.0
pyzmq==25.0.2
tqdm==4.65.0
types-tqdm==4.65.0.1
pyyaml==6.0
types-pyyaml==6.0.12.10
isort==5.12.0
black==23.3.0
pylint==2.17.4
mypy==1.3.0
